moving filter menu in own file
