+++ client/src/management/ApplicationsOverview/ApplicationOverview.tsx
@@ -1,30 +1,17 @@
-import {
-  Group,
-  TextInput,
-  Checkbox,
-  Stack,
-  Button,
-  Menu,
-  Tooltip,
-  NumberInput,
-} from '@mantine/core'
-import { IconAdjustments, IconSearch } from '@tabler/icons-react'
+import { Group, TextInput, Stack, Button, Menu, Tooltip } from '@mantine/core'
+import { IconSearch } from '@tabler/icons-react'
 import { useState, useMemo, useEffect } from 'react'
 import { TechnicalChallengeAssessmentModal } from './components/TechnicalChallengeAssessmentModal'
 import { ApplicationDatatable } from './components/ApplicationDatatable'
 import { MatchingResultsUploadModal } from './components/MatchingResultsUploadModal'
 import { useApplicationStore } from '../../state/zustand/useApplicationStore'
 import { useQuery } from '@tanstack/react-query'
-import {
-  Application,
-  ApplicationStatus,
-  ApplicationType,
-  Gender,
-} from '../../interface/application'
+import { Application, ApplicationType, Gender } from '../../interface/application'
 import { Query } from '../../state/query'
 import { getApplications } from '../../network/application'
 import { useCourseIterationStore } from '../../state/zustand/useCourseIterationStore'
 import { FilterChips } from './components/FilterChips'
+import { FilterMenu } from './components/FilterMenu'
 
 export interface Filters {
   gender: Gender[]
@@ -61,17 +48,6 @@ export const StudentApplicationOverview = (): JSX.Element => {
     },
     applicationType: [ApplicationType.DEVELOPER],
   })
-
-  const applicationStatusOptions = Object.values(ApplicationStatus)
-
-  const handleStatusFilterChange = (status: string, checked: boolean) => {
-    setFilters((currFilters) => ({
-      ...currFilters,
-      status: checked
-        ? [...currFilters.status, status]
-        : currFilters.status.filter((existingStatus) => existingStatus !== status.toString()),
-    }))
-  }
 
   const { data: fetchedDeveloperApplications, isLoading: isLoadingDeveloperApplications } =
     useQuery<Application[]>({
@@ -185,182 +161,8 @@ export const StudentApplicationOverview = (): JSX.Element => {
             setSearchQuery(e.currentTarget.value)
           }}
         />
-        <Menu withArrow closeOnItemClick={false}>
-          <Menu.Target>
-            <IconAdjustments />
-          </Menu.Target>
-          <Menu.Dropdown>
-            <Menu.Label>Gender</Menu.Label>
-            <Menu.Item>
-              <Checkbox
-                label='Male'
-                checked={filters.gender.includes(Gender.MALE)}
-                onChange={(e) => {
-                  setFilters((currFilters) => ({
-                    ...currFilters,
-                    gender: e.currentTarget.checked
-                      ? [...currFilters.gender, Gender.MALE]
-                      : currFilters.gender.filter((gender) => gender !== Gender.MALE),
-                  }))
-                }}
-              />
-            </Menu.Item>
-            <Menu.Item>
-              <Checkbox
-                label='Female'
-                checked={filters.gender.includes(Gender.FEMALE)}
-                onChange={(e) => {
-                  setFilters((currFilters) => ({
-                    ...currFilters,
-                    gender: e.currentTarget.checked
-                      ? [...currFilters.gender, Gender.FEMALE]
-                      : currFilters.gender.filter((gender) => gender !== Gender.FEMALE),
-                  }))
-                }}
-              />
-            </Menu.Item>
-            <Menu.Item>
-              <Checkbox
-                label='Unkown / Other'
-                checked={
-                  filters.gender.includes(Gender.PREFER_NOT_TO_SAY) ||
-                  filters.gender.includes(Gender.OTHER)
-                }
-                onChange={(e) => {
-                  setFilters((currFilters) => ({
-                    ...currFilters,
-                    gender: e.currentTarget.checked
-                      ? [...currFilters.gender, Gender.OTHER]
-                      : currFilters.gender.filter((gender) => gender !== Gender.OTHER),
-                  }))
-                }}
-              />
-            </Menu.Item>
 
-            <Menu.Divider />
-            <Menu.Label>Assessment</Menu.Label>
-            <Menu.Item>
-              <Checkbox
-                label='No Score'
-                checked={filters.assessment.noScore}
-                onChange={(e) => {
-                  setFilters((oldFilters: Filters) => {
-                    return {
-                      ...oldFilters,
-                      assessment: {
-                        minScore: 0,
-                        maxScore: 100,
-                        noScore: e.currentTarget.checked,
-                      },
-                    }
-                  })
-                }}
-              />
-            </Menu.Item>
-            <Menu.Item>
-              <div style={{ display: 'flex', gap: '10px', alignItems: 'center' }}>
-                <NumberInput
-                  description='Min Score'
-                  placeholder='0'
-                  value={filters.assessment.minScore}
-                  min={0}
-                  max={100}
-                  onChange={(value) => {
-                    setFilters((oldFilters: Filters) => {
-                      return {
-                        ...oldFilters,
-                        assessment: {
-                          minScore: +value,
-                          noScore: +value > 0 ? false : oldFilters.assessment.noScore,
-                          maxScore: oldFilters.assessment.noScore
-                            ? 100
-                            : oldFilters.assessment.maxScore,
-                        },
-                      }
-                    })
-                  }}
-                  style={{
-                    width: '6rem',
-                  }}
-                  styles={{
-                    input: {
-                      color: filters.assessment.minScore === 0 ? 'darkgray' : 'black',
-                      opacity: filters.assessment.minScore === 0 ? 0.8 : 1,
-                    },
-                    description: {
-                      color: 'black',
-                    },
-                  }}
-                />
-                <NumberInput
-                  description='Max Score'
-                  placeholder='100'
-                  value={filters.assessment.maxScore}
-                  min={0}
-                  max={100}
-                  onChange={(value) => {
-                    setFilters((oldFilters: Filters) => {
-                      return {
-                        ...oldFilters,
-                        assessment: {
-                          ...oldFilters.assessment,
-                          maxScore: +value,
-                          noScore: +value > 0 ? false : oldFilters.assessment.noScore,
-                        },
-                      }
-                    })
-                  }}
-                  style={{ width: '6rem' }}
-                  styles={{
-                    input: {
-                      color: filters.assessment.maxScore === 100 ? 'darkgray' : 'black',
-                      opacity: filters.assessment.maxScore === 100 ? 0.8 : 1,
-                    },
-                    description: {
-                      color: 'black',
-                    },
-                  }}
-                />
-              </div>
-            </Menu.Item>
-
-            <Menu.Divider />
-            <Menu.Label>Assessment</Menu.Label>
-            {Object.keys(ApplicationStatus).map((status) => (
-              <Menu.Item key={status}>
-                <Checkbox
-                  label={ApplicationStatus[status]}
-                  checked={filters.status.includes(status)}
-                  onChange={(e) => handleStatusFilterChange(status, e.currentTarget.checked)}
-                />
-              </Menu.Item>
-            ))}
-
-            <Menu.Divider />
-            <Menu.Item>
-              <Button
-                variant='light'
-                fullWidth
-                onClick={() => {
-                  setFilters((oldFilters: Filters) => {
-                    return {
-                      ...oldFilters,
-                      gender: [],
-                      status: [],
-                      assessment: {
-                        noScore: false,
-                        minScore: 0,
-                        maxScore: 100,
-                      },
-                    }
-                  })
-                }}
-              >
-                Reset Filters
-              </Button>
-            </Menu.Item>
-          </Menu.Dropdown>
-        </Menu>
+        <FilterMenu filters={filters} setFilters={setFilters} />
 
         <Menu withArrow>
           <Menu.Target>
